Centralize env config
