+++ src/auto/socials/mastodon_client.py
@@ -1,49 +1,9 @@
 from mastodon import Mastodon
-from dotenv import load_dotenv, find_dotenv
-from pathlib import Path
 import logging
-import os
 
-# Resolve the repository root so we can load the correct .env file
-dotenv_path = find_dotenv()
-BASE_DIR = (
-    Path(dotenv_path).resolve().parent
-    if dotenv_path
-    else Path(__file__).resolve().parents[3]
-)
-
-
-def get_mastodon_instance() -> str:
-    """Return the Mastodon instance URL from the environment."""
-    load_dotenv(BASE_DIR / ".env")
-    return os.getenv("MASTODON_INSTANCE", "https://mastodon.social")
-
-
-def get_mastodon_token() -> str | None:
-    """Return the Mastodon access token from the environment."""
-    load_dotenv(BASE_DIR / ".env")
-    return os.getenv("MASTODON_TOKEN")
+from ..config import get_mastodon_instance, get_mastodon_token
 
 logger = logging.getLogger(__name__)
-
-
-def _load_env() -> None:
-    """Load environment variables from the nearest ``.env`` file."""
-    dotenv_path = find_dotenv()
-    if dotenv_path:
-        load_dotenv(dotenv_path)
-
-
-def get_instance() -> str:
-    """Return the Mastodon instance URL."""
-    _load_env()
-    return os.getenv("MASTODON_INSTANCE", "https://mastodon.social")
-
-
-def get_access_token() -> str:
-    """Return the Mastodon access token."""
-    _load_env()
-    return os.getenv("MASTODON_TOKEN")
 
 
 def post_to_mastodon(status: str, visibility: str = "private") -> None:
